Added repositories and services, implemented some methods in services

+++ api/baytree_mentoring/src/main/java/com/baytree_mentoring/baytree_mentoring/models/Mentee.java
@@ -14,7 +14,7 @@ import javax.validation.constraints.NotNull;
 
 public class Mentee {
     @Id
-    private int menteeId;
+    private double menteeId;
 
     @NotNull
     private String firstName;
