+++ api/baytree_mentoring/src/main/java/com/baytree_mentoring/baytree_mentoring/models/Mentor.java
@@ -15,10 +15,9 @@ import java.util.List;
 @Entity
 @Getter
 @AllArgsConstructor
-
 public class Mentor {
     @Id
-    private int mentorId;
+    private long mentorId;
 
     @NotNull
     private String firstName;
